+++ clients/fortnite_tracker.py
@@ -8,7 +8,7 @@ from urllib.parse import unquote
 import aiohttp
 from bs4 import BeautifulSoup
 
-import clients.discord_base as discord_base
+import utils.discord as discord_utils
 from database.mysql import MySQL
 from exceptions import UserDoesNotExist, NoSeasonDataError
 from utils.dates import get_playing_session_date
@@ -189,9 +189,9 @@ def _find_mode_stat(stat_name, mode_stats):
 
 def _create_message(username, stats_breakdown):
     """ Create player stats Discord message """
-    return discord_base.create_stats_message(
+    return discord_utils.create_stats_message(
         title=f"Username: {username}",
-        desc=discord_base.create_wins_str(stats_breakdown["all"]),
+        desc=discord_utils.create_wins_str(stats_breakdown["all"]),
         color_metric=stats_breakdown["all"]["KD"],
         create_stats_func=_create_stats_str,
         stats_breakdown=stats_breakdown,
@@ -203,8 +203,8 @@ def _create_stats_str(mode, stats_breakdown):
     """ Create stats string for output """
     mode_stats = stats_breakdown[mode]
     return (f"KD: {mode_stats['KD']} • "
-            f"Wins: {int(mode_stats['Top1']):,} • "
-            f"Win Percentage: {mode_stats['WinRatio']:,.1f}% • "
+            f"Wins: {int(mode_stats['Wins']):,} • "
+            f"Win Percentage: {mode_stats['Win Percentage']:,.1f}% • "
             f"Matches: {int(mode_stats['Matches']):,} • "
             f"TRN: {int(mode_stats['TRNRating']):,}")
 
@@ -219,8 +219,8 @@ async def _track_player(username, stats_breakdown):
             "mode": mode,
             "kd": stats["KD"],
             "games": stats["Matches"],
-            "wins": stats["Top1"],
-            "win_rate": stats["WinRatio"],
+            "wins": stats["Wins"],
+            "win_rate": stats["Win Percentage"],
             "trn": stats["TRNRating"],
             "date_added": get_playing_session_date()
         })
