+++ clients/stats.py
@@ -1,13 +1,13 @@
 from collections import defaultdict
 
-import clients.discord_base as discord_base
+import utils.discord as discord_utils
 from database.mysql import MySQL
 
 
 async def send_stats_diff_today(ctx, username):
     """ Sends the stats diff between today and the last play date """
     mysql = await MySQL.create()
-    season_id = discord_base.get_season_id()
+    season_id = discord_utils.get_season_id()
 
     player_snapshots = await mysql.fetch_player_stats_diff_today(
         username,
@@ -40,17 +40,18 @@ def _breakdown_player_snapshots(player_snapshots):
         row_current = processed_snapshots["current"][mode]
         row_previous = processed_snapshots["previous"].get(mode, defaultdict(int))
 
-        # TODO: Change stats to our own naming convention
+        # TODO: Revert the key name changes to Fortnite Tracker format
+
         stats[mode] = {
             "KD": {
                 "current": row_current.get("kd", 0),
                 "diff": _pad_symbol(f'{row_current.get("kd", 0) - row_previous.get("kd", 0):,.2f}')
             },
-            "Top1": {
+            "Wins": {
                 "current": row_current.get("wins", 0),
                 "diff": _pad_symbol(f'{row_current.get("wins", 0) - row_previous.get("wins", 0):,}')
             },
-            "WinRatio": {
+            "Win Percentage": {
                 "current": row_current.get("win_rate", 0),
                 "diff": _pad_symbol(f'{row_current.get("win_rate", 0) - row_previous.get("win_rate", 0):,.1f}')
             },
@@ -74,7 +75,7 @@ def _pad_symbol(val: str):
 
 def _create_stats_diff_message(username, stats_breakdown):
     """ Create opponent stats Discord message """
-    return discord_base.create_stats_message(
+    return discord_utils.create_stats_message(
         title=f"Username: {username}",
         desc=_create_wins_diff_str(stats_breakdown["all"]),
         color_metric=stats_breakdown["all"]["KD"]["current"],
@@ -86,7 +87,7 @@ def _create_stats_diff_message(username, stats_breakdown):
 
 def _create_wins_diff_str(win_stats):
     """ Create stats string for output """
-    wins_str = f"{int(win_stats['Top1']['current'])} ({win_stats['Top1']['diff']})"
+    wins_str = f"{int(win_stats['Wins']['current'])} ({win_stats['Wins']['diff']})"
     matches_str = f"{int(win_stats['Matches']['current']):,} ({win_stats['Matches']['diff']}) played"
     return f"Wins: {wins_str} / {matches_str}"
 
@@ -95,8 +96,8 @@ def _create_stats_diff_str(mode, stats_breakdown):
     """ Create stats string for output """
     mode_stats = stats_breakdown[mode]
     return (f"KD: {mode_stats['KD']['current']} ({mode_stats['KD']['diff']}) • "
-            f"Wins: {int(mode_stats['Top1']['current'])} ({mode_stats['Top1']['diff']}) • "
-            f"Win Percentage: {mode_stats['WinRatio']['current']:,.1f}% ({mode_stats['WinRatio']['diff']}%) • "
+            f"Wins: {int(mode_stats['Wins']['current'])} ({mode_stats['Wins']['diff']}) • "
+            f"Win Percentage: {mode_stats['Win Percentage']['current']:,.1f}% ({mode_stats['Win Percentage']['diff']}%) • "
             f"Matches: {int(mode_stats['Matches']['current'])} ({mode_stats['Matches']['diff']}) • "
             f"TRN: {int(mode_stats['TRNRating']['current'])} ({mode_stats['TRNRating']['diff']})")
 
@@ -137,9 +138,9 @@ def _breakdown_opponent_average_stats(opponent_avg_stats):
 
 def _create_opponents_stats_message(opponent_stats_breakdown):
     """ Create opponent stats Discord message """
-    return discord_base.create_stats_message(
+    return discord_utils.create_stats_message(
         title="Opponent Average Stats Today",
-        desc=discord_base.create_wins_str(opponent_stats_breakdown["all"]),
+        desc=discord_utils.create_wins_str(opponent_stats_breakdown["all"]),
         color_metric=opponent_stats_breakdown["all"]["KD"],
         create_stats_func=_create_opponent_stats_str,
         stats_breakdown=opponent_stats_breakdown
@@ -163,6 +164,6 @@ async def rate_opponent_stats_today(ctx):
 
     opponent_stats_breakdown = _breakdown_opponent_average_stats(opponent_stats)
 
-    skill_rate = discord_base.calculate_skill_rate_indicator(
+    skill_rate = discord_utils.calculate_skill_rate_indicator(
         opponent_stats_breakdown["all"]["KD"])
     await ctx.send(skill_rate)
